Sets a basic way of dealing with #48
Splits the logging in two loggers (console and file) so the logs can be formatted separately.

A more complete console logging is required though. Right now it only should the message, but it would be nice to include also the parametters in a human-redable and friendly way.

Fixing it using a single Logger will also be nice, but not worth spending too much on.